+++ src/main/java/kr/hhplus/be/server/persistence/balance/BalanceJpaRepository.java
@@ -1,11 +1,10 @@
-package kr.hhplus.be.server.persistence.balance.repository;
+package kr.hhplus.be.server.persistence.balance;
 
 import kr.hhplus.be.server.domain.balance.Balance;
-import kr.hhplus.be.server.persistence.balance.entity.BalanceJpaEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BalanceJpaRepository extends JpaRepository<Balance,Long> {
 
-    BalanceJpaEntity findByUserId(long userId);
+    Balance findByUserId(long userId);
 
 }
